refactor(util): UTF-8 デコーダを utf8_codec.h に集約
document_utils.cpp と doc_dwrite_bridge.cpp に分散していた UTF-8 マルチ
バイト decode を src/util/utf8_codec.h に統合。UTF-8 / UTF-16 両対応で
DecodeAt / SnapToCpStart / DecodePrev を提供し、不正バイト・truncated・
不正 continuation はすべて U+FFFD に正規化する。

副次効果として GenerateAnchorIdInto と WideViewForDWrite で不正
continuation byte の検証が新たに有効化され、ゴミ code point の混入を
防止する。
